% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{SiChgcar}
\alias{SiChgcar}
\title{Output of the \code{\link{vasp.import}} function applied on the CHGCAR file of a single silicium atom in a fcc cell.}
\format{Data frame containing four columns
\itemize{
  \item{ charge: Consists of four columns: "x", "y", "z" containing the Cartesian coordinates to a grid point and "charge" containing the charge density on this specific grid point.}
  \item{ atoms: Consists of four columns: "x", "y", "z" containing the Cartesian coordinates to the atoms within the cell and "type" providing a factor to distinguish the individual types of atoms.}
  \item{ lattice:  Consists of three columns: "x", "y", "z" containing the components of the three lattice vectors of the unit cell.}
}}
\description{
Output of the \code{\link{vasp.import}} function applied on the CHGCAR file of a single silicium atom in a fcc cell.
}

