% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vasp-import.R
\name{vasp.import}
\alias{vasp.import}
\title{vasp.import}
\usage{
vasp.import(path)
}
\arguments{
\item{path}{Character vector containing the path to the CHGCAR file.}
}
\value{
List of three named data frames:
\itemize{
  \item{ charge: Consists of four columns: "x", "y", "z" containing the Cartesian coordinates to a grid point and "charge" containing the charge density on this specific grid point.}
  \item{ atoms: Consists of four columns: "x", "y", "z" containing the Cartesian coordinates to the atoms within the cell and "type" providing a factor to distinguish the individual types of atoms.}
  \item{ lattice:  Consists of three columns: "x", "y", "z" containing the components of the three lattice vectors of the unit cell.}
 }
}
\description{
Imports the CHGCAR file of a VASP calculation into R. Should also run with the CHG file.
}
\details{
The CHGCAR file got a very odd format which has to be reformatted to ensure the its compatibility with R. In the CHGCAR file the number of grid points of the specific axes are given in the 12 column. In the charge data the x coordinate is the fastest and the z the slowest varying index. The charge density is also divided by the volume of the cell inside of this script (VASP internal stuff). The individual lattice vectors in the are given in different rows and their x, y and z are provided in the different columns.
}
\author{
Philipp Mueller
}
\seealso{
Other vasp: \code{\link{vasp.bonds}};
  \code{\link{vasp.reproduce}};
  \code{\link{vasp.rotate.cell}}
}

