% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{vasp.plane}
\alias{vasp.plane}
\title{Extracts a plane of charge orthogonal to the z axis.}
\usage{
vasp.plane(x, height = mean(x$atoms[x$atoms$type == unique(x$atoms$type)[1],
  ]$z))
}
\arguments{
\item{x}{Object result of class "vasp".}

\item{height}{Position on the z axis at which the slice through the charge density is going to be performed. The default value is the mean z value of all atoms of the first type listed in x$atoms$type (since the package was originally intended for layered materials).}
}
\value{
Object of class vasp containing just one z plane in its $charge element.
}
\details{
Only the charge density is sliced.
}
\author{
Philipp Mueller
}
\seealso{
Other vasp: \code{\link{vasp.bonds}},
  \code{\link{vasp.diff}}, \code{\link{vasp.import}},
  \code{\link{vasp.reproduce}},
  \code{\link{vasp.rotate.cell}}
}

