% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{vasp.reproduce}
\alias{vasp.reproduce}
\title{vasp.reproduce}
\usage{
vasp.reproduce(lattice, cell, v1 = NULL, v2 = NULL, v3 = NULL)
}
\arguments{
\item{lattice}{Data frame of dimension 3x3 containing the three lattice vectors in separate rows and their coordinates in different columns. See \code{\link{vasp.import}}.}

\item{cell}{Data frame containing at least two named columns "x" and "y" containing Cartesian coordinates.}

\item{v1}{Number of reproductions of the first lattice vector of the unit cell given in a sequence.}

\item{v2}{Number of reproductions of the second lattice vector of the unit cell given in a sequence.}

\item{v3}{Number of reproductions of the third lattice vector of the unit cell given in a sequence.}
}
\value{
Input 'cell' extended by the additional content of the reproduced unit cell.
}
\description{
Reproduce the content of a unit cell in multiple direction of its lattice vectors.
}
\details{
The number of reproductions has to be provided as a sequence. E.g. seq(-3,3)
}
\author{
Philipp Mueller
}
\seealso{
Other vasp: \code{\link{vasp.bonds}};
  \code{\link{vasp.import}}; \code{\link{vasp.rotate.cell}}
}

