% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{vasp.rotate.cell}
\alias{vasp.rotate.cell}
\title{vasp.rotate.cell}
\usage{
vasp.rotate.cell(lattice, cell, angle = NULL)
}
\arguments{
\item{lattice}{Data frame of dimension 3x3 containing the three lattice vectors in separate rows and their coordinates in different columns. See \code{\link{vasp.import}}.}

\item{cell}{Data frame containing at least two named columns "x" and "y" containing Cartesian coordinates.}

\item{angle}{Specifies the angle with which the cell is going to be rotated. If no value is supplied the angle between the two lattice vectors is determined and the content will be shifted by half of this angle in the positive direction.}
}
\value{
Same object like the input "cell" but with modified coordinates.
}
\description{
Rotates the content of a cell with a specific angle.
}
\details{
In this version of the function only a rotation in the x-y plane is possible. The supplied content of the cell has to consist of named columns "x" and "y" for the Cartesian coordinates. If no angle is provided the angle between the two lattice vectors is determined and the content will be shifted by half of this angle in the positive direction. A rotation matrix is applied to all the coordinates in the cell using a cpp loop.
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{vasp.import}}

Other vasp: \code{\link{vasp.bonds}};
  \code{\link{vasp.import}}; \code{\link{vasp.reproduce}}
}

