% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{vasp.bonds}
\alias{vasp.bonds}
\title{vasp.bonds}
\usage{
vasp.bonds(atoms, distance)
}
\arguments{
\item{atoms}{Data frame containing the x, y and z position of the atoms in cartesian coordinates as named columns and a fourth column names "type" containing a character specifying the atom species. See \code{\link{vasp.import}}$atoms.}

\item{distance}{Vector of length unique( atoms$type ) containing the bonding distances between the individual atoms. To establish a bond two atoms must be closer than this value. The order in this distances vector has to be the same as in unique( atoms$type ).}
}
\value{
Data frame with seven named columns "x.begin", "y.begin", "z.begin", "x.end", "y.end", "z.end" and "type" containing the start and end points of the bonds in the system as well as the type of atoms between which the bond is established. The later is necessary for the correct coloring and can be easily extended to bonds of many different atom types.
}
\description{
Generates a data frame containing the start and the endpoints of all bounds in the system.
}
\details{
Two atoms share a bond if they a closer than a specified distance. Since this function was originally written for a layered system the bonds are only created within one type of atoms. But its quite easily extendible to bonds between several kinds of atoms.
}
\author{
Philipp Mueller
}
\seealso{
Other vasp: \code{\link{vasp.import}};
  \code{\link{vasp.reproduce}};
  \code{\link{vasp.rotate.cell}}
}

