% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{vasp.reproduce}
\alias{vasp.reproduce}
\title{Reproduce the content of a unit cell in multiple direction of its lattice vectors.}
\usage{
vasp.reproduce(x, lattice = NULL, x.rep = NULL, y.rep = NULL,
  z.rep = NULL, x.window = NULL, y.window = NULL, z.window = NULL)
}
\arguments{
\item{x}{A data.frame containing at least two named columns "x" and "y" containing Cartesian coordinates or an object of class "vasp" to reproduce its atom positions and charge in one step.}

\item{lattice}{A data.frame of dimension 3x3 containing the three lattice vectors in separate rows and their coordinates in different columns. See \code{\link{vasp.import}}. If an object of class "vasp" was supplied in the 'x' argument 'lattice' can be omitted.}

\item{x.rep}{Number of reproductions along the first lattice vector of the unit cell given in a sequence.}

\item{y.rep}{Number of reproductions along the second lattice vector of the unit cell given in a sequence.}

\item{z.rep}{Number of reproductions along the third lattice vector of the unit cell given in a sequence.}
}
\value{
Input 'x' extended by the additional content of the reproduced unit cell.
}
\details{
The number of reproductions has to be provided as a numerical vector (e.g. seq(-3,3)). Generic function working with both the specific content of the x, like the atom positions, and an object of class "vasp".
}
\author{
Philipp Mueller
}
\seealso{
Other vasp: \code{\link{vasp.bonds}},
  \code{\link{vasp.diff}}, \code{\link{vasp.import}},
  \code{\link{vasp.plane}}, \code{\link{vasp.rotate.cell}}
}

