% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vasp-utils.R
\name{vasp.rotate.cell}
\alias{vasp.rotate.cell}
\title{Rotates the content of a cell by a specific angle.}
\usage{
vasp.rotate.cell(cell, lattice = NULL, angle = NULL)
}
\arguments{
\item{cell}{A data.frame containing at least two named columns "x" and "y" with Cartesian coordinates or an object of class "vasp".}

\item{lattice}{A data.frame of dimension 3x3 containing the lattice vectors in its rows and their coordinates in its columns. See \code{\link{vasp.import}}. If the argument 'cell' is an object of class "vasp" 'lattice' can is omitted.}

\item{angle}{Specifies the angle with which the cell is going to be rotated. If no value is supplied the angle between the two lattice vectors is determined and the content will be shifted clockwise by half of its value.}
}
\value{
Same object like the input "cell" but with modified coordinates.
}
\details{
In the current version of this function only a rotation in the x-y plane is possible. If no angle is provided the angle between the two lattice vectors is determined and the content will be shifted clockwise by half of its value. Internally a rotation matrix is applied to all the coordinates in the cell. It is intended to be a generic function working with both the specific content of the cell, like the atom positions, and an object of class "vasp".
}
\author{
Philipp Mueller
}
\seealso{
\code{\link{vasp.import}}

Other vasp: \code{\link{vasp.bonds}},
  \code{\link{vasp.diff}}, \code{\link{vasp.import}},
  \code{\link{vasp.reproduce}}
}

